+++ js/argument.js
@@ -1,0 +1,13 @@
+const generatedObject = 25;
+const maxIdComment = 500;
+const minLikes = 15;
+const maxLikes = 200;
+const minComment = 0;
+const maxComment = 30;
+
+export {generatedObject};
+export {maxIdComment};
+export {minLikes};
+export {maxLikes};
+export {minComment};
+export {maxComment};
