+++ js/massive.js
@@ -1,0 +1,55 @@
+const DESCRIPTION = [
+  'На пляже',
+  'Метро',
+  'Прекрасный день',
+  'Авто',
+  'Коты',
+  'Собака',
+  'Домашние любимцы',
+  'Ужасный день',
+  'Завтрак',
+  'Обед',
+  'Обед в ресторане',
+  'Ужин',
+  'Праздничный ужин',
+  'Пасмурно',
+  'Дождливый день',
+  'Длинная дорога',
+  'Путешествие',
+  'Любимая улица',
+  'Пришли зима',
+  'Пришла осень',
+  'Пришло лето',
+  'Пришла весна',
+  'Работа',
+  'Учеба',
+  'Пятничный отдых',
+];
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const NAME = [
+  'Дмитрий',
+  'Александр',
+  'Геннадий',
+  'Вадим',
+  'Артем',
+  'Яна',
+  'Алексей',
+  'Жана',
+  'Сергей',
+  'Олег',
+  'Наталья',
+  'Юлия',
+  'Анна',
+  'Екатерина',
+];
+
+export {DESCRIPTION};
+export {MESSAGE};
+export {NAME};
